crud for note and add validation

+++ src/main/java/lk/ijse/noteCollector/imp/UserServiceIMPL.java
@@ -1,8 +1,12 @@
 package lk.ijse.noteCollector.imp;
 
+import lk.ijse.noteCollector.customStatusCode.SelectedUserStatus;
 import lk.ijse.noteCollector.dao.UserDao;
+import lk.ijse.noteCollector.dto.UserStatus;
 import lk.ijse.noteCollector.dto.impl.UserDTO;
 import lk.ijse.noteCollector.entity.impl.UserEntity;
+import lk.ijse.noteCollector.exception.DataPersistException;
+import lk.ijse.noteCollector.exception.UserNotFoundException;
 import lk.ijse.noteCollector.service.UserService;
 import lk.ijse.noteCollector.util.Mapping;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +14,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
+
 @Service
 @Transactional
 public class UserServiceIMPL implements UserService {
@@ -18,9 +24,12 @@ public class UserServiceIMPL implements UserService {
     @Autowired
     private Mapping mapping;
     @Override
-    public UserDTO saveUser(UserDTO userDTO) {
+    public void saveUser(UserDTO userDTO) {
         UserEntity savedUser = userDao.save(mapping.toUserEntity(userDTO));
-        return mapping.toUserDto(savedUser);
+        if(savedUser==null){
+            throw new DataPersistException("user not saved");
+        }
+//        return mapping.toUserDto(savedUser);
     }
 
     @Override
@@ -30,18 +39,37 @@ public class UserServiceIMPL implements UserService {
     }
 
     @Override
-    public UserDTO getUser(String userId) {
-     UserEntity selectedUser=  userDao.getReferenceById(userId);
-     return mapping.toUserDto(selectedUser);
+    public UserStatus getUser(String userId) {
+        if(userDao.existsById(userId)){
+            UserEntity selectedUser = userDao.getReferenceById(userId);
+            return mapping.toUserDto(selectedUser);
+        }else {
+            return new SelectedUserStatus(2, "User with id " + userId + " not found");
+        }
     }
 
     @Override
     public void deleteUser(String userId) {
-      userDao.deleteById(userId);
+        Optional<UserEntity> existedUser = userDao.findById(userId);
+        if(!existedUser.isPresent()){
+            throw new UserNotFoundException("User with id " + userId + " not found");
+        }else {
+            userDao.deleteById(userId);
+        }
+
     }
 
     @Override
-    public boolean updateUser(String noteId, UserDTO userDTO) {
-        return false;
+    public void updateUser(String userId, UserDTO userDTO) {
+
+
+        Optional<UserEntity> tmpUser = userDao.findById(userId);
+        if(tmpUser.isPresent()) {
+            tmpUser.get().setFirstName(userDTO.getFirstName());
+            tmpUser.get().setLastName(userDTO.getLastName());
+            tmpUser.get().setEmail(userDTO.getEmail());
+            tmpUser.get().setPassword(userDTO.getPassword());
+            tmpUser.get().setProfilePic(userDTO.getProfilePic());
+        }
     }
 }
